Ml2 dto (#14)
* add func dto

* separated into files
