+++ src/fml_doc_gen/write_docstring_to_file.py
@@ -1,0 +1,34 @@
+
+def write_docstring_to_file(docstring: str, output_file: str) -> None:
+    """
+    Writes the generated docstring to a specified output file.
+
+    Parameters
+    ----------
+    docstring : str
+        The docstring to be written to the file.
+    output_file : str
+        The path to the output file.
+
+    Returns
+    -------
+    None
+        This function does not return anything.
+
+    Examples
+    --------
+    >>> docstring = \"\"\"Parameters
+    ----------
+    a : int
+    b : int
+
+    Returns
+    -------
+    int
+    \"\"\"
+    >>> output_file = 'docstring_output.txt'
+    >>> write_docstring_to_file(docstring, output_file)
+    # This writes the docstring to 'docstring_output.txt'
+    """
+
+    pass
